rename utility to gdutilfunc
