+++ cmd/generate/gdutilfunc/generate.go
@@ -1,4 +1,4 @@
-package utility
+package gdutilfunc
 
 import (
 	"bytes"
@@ -53,7 +53,7 @@ func GenerateUtilityFunctions(projectPath string, extensionApi extensionapiparse
 		return err
 	}
 
-	filename := filepath.Join(projectPath, "pkg", "utility", fmt.Sprintf("utilityfunctions.gen.go"))
+	filename := filepath.Join(projectPath, "pkg", "gdutilfunc", fmt.Sprintf("utilityfunctions.gen.go"))
 
 	f, err := os.Create(filename)
