+++ cmd/generate/gdutilfunc/utilityfunctions.go.tmpl
@@ -1,5 +1,5 @@
 {{ $view := . -}}
-package utility
+package gdutilfunc
 
 /*------------------------------------------------------------------------------
 //   This code was generated by template utilityfunctions.go.tmpl.
