+++ pkg/gdutilfunc/lib.go
@@ -1,0 +1,14 @@
+package gdutilfunc
+
+/*
+#cgo CFLAGS: -I${SRCDIR}/../../godot_headers -I${SRCDIR}/../../pkg/log -I${SRCDIR}/../../pkg/gdutilfunc
+#include <godot/gdextension_interface.h>
+*/
+import "C"
+import (
+	"unsafe"
+)
+
+var (
+	nullptr = unsafe.Pointer(nil)
+)
