+++ pkg/gdutilfunc/utilityfunctions.gen.go
@@ -1,4 +1,4 @@
-package utility
+package gdutilfunc
 
 /*------------------------------------------------------------------------------
 //   This code was generated by template utilityfunctions.go.tmpl.
