--- pkg/utility/lib.go
@@ -1,14 +1,0 @@
-package utility
-
-/*
-#cgo CFLAGS: -I${SRCDIR}/../../godot_headers -I${SRCDIR}/../../pkg/log -I${SRCDIR}/../../pkg/utility
-#include <godot/gdextension_interface.h>
-*/
-import "C"
-import (
-	"unsafe"
-)
-
-var (
-	nullptr = unsafe.Pointer(nil)
-)
